requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
typer[all]==0.12.3
PyYAML==6.0.1
rich>=13.0.0
